requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
supabase
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-decouple
aiofiles